+++ atem/load.go
@@ -60,7 +60,7 @@ func LoadFromJson(src []byte) Prog {
 			fd.isMereAlias = !iscall
 		} else if len(fd.Args) >= 2 { // check if selector and set so
 			if argref, isa := fd.Body.(ExprArgRef); isa {
-				fd.selector = int(argref)
+				fd.selector = int8(argref)
 			} else if call, isc := fd.Body.(*ExprCall); isc {
 				if argref, isa = call.Callee.(ExprArgRef); isa && argref != -1 {
 					for ia := range call.Args {
@@ -69,7 +69,7 @@ func LoadFromJson(src []byte) Prog {
 						}
 					}
 					if isa {
-						fd.selector = len(call.Args)
+						fd.selector = int8(len(call.Args))
 					}
 				}
 			}
@@ -84,15 +84,21 @@ func LoadFromJson(src []byte) Prog {
 
 func (me Prog) postLoadPreProcess(expr Expr) Expr {
 	for fnr, _ := expr.(ExprFuncRef); fnr > 0 && me[fnr].isMereAlias; fnr, _ = expr.(ExprFuncRef) {
+		// main reason for this pre-reduction is to not need this check plus
+		// reduction in the interpreter loop on every `ExprFuncRef` occurrence.
+		// especially in pre-optimized input programs it would always be incurred
+		// but never "fire". to also allow less-/non-optimized input programs,
+		// we pre-reduce this on-load. another micro-optimization of sorts.
 		expr = me[fnr].Body
 	}
+	// now, detect and mark all recognizable "closures" as defined in the doc-comment for the `ExprCall.IsClosure` field
 	if call, iscall := expr.(*ExprCall); iscall {
 		call.Callee = me.postLoadPreProcess(call.Callee)
 		for i := range call.Args {
 			call.Args[i] = me.postLoadPreProcess(call.Args[i])
 		}
 		if f, _ := call.Callee.(ExprFuncRef); f > 0 {
-			diff := len(me[f].Args) - len(call.Args)
+			diff := int8(len(me[f].Args)) - int8(len(call.Args))
 			for i := 0; (diff > 0) && (i < len(call.Args)); i++ {
 				_, isa := call.Args[i].(ExprArgRef)
 				if c, isc := call.Args[i].(*ExprCall); isa || (isc && c.IsClosure == 0) {
